♻️ refactor: GameSlot 리팩토링
- 관심사 분리
  - GameSlotDetail
  - GameSlotSummary
- Recoil, Suspense를 이용한 비동기 처리
- rankToString 함수 이동
